+++ lib/kubernetes-deploy/options_helper.rb
@@ -2,19 +2,42 @@
 
 module KubernetesDeploy
   module OptionsHelper
-    def self.default_and_check_template_dir(template_dir)
-      if !template_dir && ENV.key?("ENVIRONMENT")
-        template_dir = "config/deploy/#{ENV['ENVIRONMENT']}"
+    class OptionsError < StandardError; end
+
+    STDIN_TEMP_FILE = "from_stdin.yml.erb"
+    class << self
+      def with_validated_template_dir(template_dir)
+        if template_dir == '-'
+          Dir.mktmpdir("kubernetes-deploy") do |dir|
+            template_dir_from_stdin(temp_dir: dir)
+            yield dir
+          end
+        else
+          yield default_template_dir(template_dir)
+        end
       end
 
-      if !template_dir || template_dir.empty?
-        puts "Template directory is unknown. " \
-          "Either specify --template-dir argument or set $ENVIRONMENT to use config/deploy/$ENVIRONMENT " \
-        + "as a default path."
-        exit(1)
+      private
+
+      def default_template_dir(template_dir)
+        if ENV.key?("ENVIRONMENT")
+          template_dir = File.join("config", "deploy", ENV['ENVIRONMENT'])
+        end
+
+        if !template_dir || template_dir.empty?
+          raise OptionsError, "Template directory is unknown. " \
+            "Either specify --template-dir argument or set $ENVIRONMENT to use config/deploy/$ENVIRONMENT " \
+            "as a default path."
+        end
+
+        template_dir
       end
 
-      template_dir
+      def template_dir_from_stdin(temp_dir:)
+        File.open(File.join(temp_dir, STDIN_TEMP_FILE), 'w+') { |f| f.print($stdin.read) }
+      rescue IOError, Errno::ENOENT => e
+        raise OptionsError, e.message
+      end
     end
   end
 end